Split out fractal settings table from gtkfractal.T
